+++ config/medium/TheCashRegister/code/CGCode.js
@@ -1,0 +1,7 @@
+const register = readline();
+const goalAmount = parseInt(readline());
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('list of coins');