+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Markov ants".
 - configuration for "One for his Nobs".
 - configuration for "Coastline".
+- configuration for "The Cash Register".
 
 ## [1.6.0] - 2024-09-02
 ### Added